Create seperate file to init firebase data
